IFC2x3 geolocation: properly handle true north
In IFC2x3, true north is given as a vector in the world coordinate system, as
are all coordinates. Any rotation of the WCS with regard to a larger context
(formerly extracted mistakenly as "projectNorth") is not relevant.

We need to rotate everything from the WCS y-axis to the true north vector.